moved scoring methods to separate package
