+++ sciencebeam_judge/evaluation/scoring_types/list.py
@@ -12,7 +12,7 @@ from ..math import safe_mean
 
 from ..normalization import normalize_string
 
-from ..scoring_methods import ScoringMethod, get_scoring_methods
+from ..scoring_methods.scoring_methods import ScoringMethod, get_scoring_methods
 
 from ..match_scoring import (
     get_match_score_obj_for_score_fn,
