+++ tests/evaluation_pipeline_test.py
@@ -29,7 +29,7 @@ from sciencebeam_judge.evaluation_pipeline import (
 )
 
 from sciencebeam_judge.evaluation.match_scoring import MatchScoringProps
-from sciencebeam_judge.evaluation.scoring_methods import ScoringMethodNames
+from sciencebeam_judge.evaluation.scoring_methods.scoring_methods import ScoringMethodNames
 from sciencebeam_judge.evaluation.scoring_types.scoring_types import ScoringTypeNames
 from sciencebeam_judge.evaluation.document_scoring import DocumentScoringProps
 from sciencebeam_judge.evaluation.score_aggregation import SummaryScoresProps
